requirements.txt
django==2.1.7
orbit-predictor==1.8.6
attrs==18.2.0
pytz
python-dateutil
djangorestframework
